+++ project/app/Http/Livewire/Tasks/Month.php
@@ -6,6 +6,7 @@ use App\Http\Livewire\Properties\TasksProperty;
 use App\Task;
 use Carbon\Carbon;
 use Carbon\Translator;
+use Illuminate\Support\Str;
 use Livewire\Component;
 
 class Month extends Component
@@ -19,8 +20,9 @@ class Month extends Component
     public function mount()
     {
         $this->month = Carbon::now()->month;
-        $carbon      = Carbon::now()->setMonth($this->month)->locale('fr_FR');
-        $this->title = "Tickets terminés | {$carbon->getTranslatedMonthName()} {$carbon->year}";
+        $carbon = Carbon::now()->setMonth($this->month)->locale('fr_FR');
+        $monthName = Str::ucfirst($carbon->getTranslatedMonthName());
+        $this->title = "{$monthName} {$carbon->year}";
     }
 
     public function nextMonth()
@@ -40,13 +42,14 @@ class Month extends Component
 
     public function getTasksProperty() : TasksProperty
     {
-        $carbon      = Carbon::now()->setMonth($this->month)->locale('fr_FR');
-        $this->title = "Tickets terminés | {$carbon->getTranslatedMonthName()} {$carbon->year}";
-        $from        = $carbon->startOfMonth()->format('Y-m-d');
-        $to          = $carbon->endOfMonth()->format('Y-m-d');
-        $tasks       = Task::where('day', '>=', $from)->where('day', '<=', $to);
-        $all         = $tasks->count();
-        $done        = $tasks->where('done', true)->count();
+        $carbon = Carbon::now()->setMonth($this->month)->locale('fr_FR');
+        $monthName = Str::ucfirst($carbon->getTranslatedMonthName());
+        $this->title = "{$monthName} {$carbon->year}";
+        $from = $carbon->startOfMonth()->format('Y-m-d');
+        $to = $carbon->endOfMonth()->format('Y-m-d');
+        $tasks = Task::where('day', '>=', $from)->where('day', '<=', $to);
+        $all = $tasks->sum('estimation');
+        $done = $tasks->where('done', true)->sum('estimation');
 
         return new TasksProperty($done, $all);
     }